Refactoring runners
