+++ stepik-union/src/main/java/org/stepik/core/SupportedLanguages.kt
@@ -35,10 +35,10 @@ enum class SupportedLanguages constructor(val title: String,
             "Main.hs",
             "--",
             "module Main where\nmain = print $ \"Hello, world!\""),
-    JAVA8("Java 8", "java8", "Main.java", "//", "class Main {", "}", testRunner = JavaTestRunner),
+    JAVA8("Java 8", "java8", "Main.java", "//", "class Main {", "}", testRunner = JavaTestRunner()),
     JAVA("Java", "java", "Main.java", "//", "class Main {", "}", JAVA8),
     JAVASCRIPT("JavaScript", "javascript", "main.js", "//"),
-    KOTLIN("Kotlin", "kotlin", "Main.kt", "//", testRunner = KotlinTestRunner),
+    KOTLIN("Kotlin", "kotlin", "Main.kt", "//", testRunner = KotlinTestRunner()),
     MONO_CS("Mono c#", "mono c#", "main.cs", "//"),
     OCTAVE("Octave", "octave", "main.m", "%"),
     PASCAL_ABC("PascalABC.NET", "pascalabc", "Main.pas", "//"),
